back-end/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Babel==2.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-RESTX==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
marshmallow==3.20.1
Pillow==10.1.0
python-magic==0.4.27
boto3==1.34.0
redis==5.0.1
pytest==7.4.3
black==23.12.1
flake8==7.0.0
